backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
eth-keys==0.4.0
eth-utils==2.3.1
eth-account==0.10.0
hyperliquid-python-sdk==0.6.0
requests==2.31.0
urllib3==2.1.0
web3==6.15.1
Werkzeug==3.0.1
